HOTT-3746: PR feedback
