+++ app/mailers/application_mailer.rb
@@ -1,0 +1,3 @@
+class ApplicationMailer < ActionMailer::Base
+  default from: TradeTariffBackend.from_email
+end
