+++ app/views/exchange_rates_mailer/monthly_files.html.erb
@@ -1,6 +1,6 @@
 <p>
   Please use the following links to download the exchange rate files for <%= @month_and_year %>.<br>
-  Links: <a href="<%= TradeTariffBackend.frontend_host + @csv_hmrc.file_path %>">CSV HMRC</a> and <a href="<%= TradeTariffBackend.frontend_host + @xml.file_path %>">XML</a>
+  Links: <a href="<%= TradeTariffBackend.frontend_host + @csv_hmrc.file_path %>">CSV HMRC link</a> and <a href="<%= TradeTariffBackend.frontend_host + @xml.file_path %>">XML link</a>
 </p>
 
 <p>
