+++ spec/services/exchange_rates/monthly_exchange_rates_service_spec.rb
@@ -17,7 +17,7 @@ RSpec.describe ExchangeRates::MonthlyExchangeRatesService do
         allow(ExchangeRates::UpdateCurrencyRatesService).to receive(:new).and_return(update_service)
         allow(ExchangeRates::UploadMonthlyFileService).to receive(:call)
         allow(SlackNotifierService).to receive(:call)
-        allow(ExchangeRates::Mailer).to receive(:monthly_files)
+        allow(ExchangeRatesMailer).to receive(:monthly_files)
 
         call_service
       end
@@ -27,7 +27,7 @@ RSpec.describe ExchangeRates::MonthlyExchangeRatesService do
       it { expect(ExchangeRates::UploadMonthlyFileService).to have_received(:call).with(:monthly_xml) }
       it { expect(ExchangeRates::UploadMonthlyFileService).to have_received(:call).with(:monthly_csv_hmrc) }
       it { expect(SlackNotifierService).to have_received(:call).with(/Exchange rates for the current month/) }
-      it { expect(ExchangeRates::Mailer).to have_received(:monthly_files) }
+      it { expect(ExchangeRatesMailer).to have_received(:monthly_files) }
     end
 
     context 'when tomorrow is not penultimate Thursday' do
@@ -37,7 +37,7 @@ RSpec.describe ExchangeRates::MonthlyExchangeRatesService do
         allow(ExchangeRates::UpdateCurrencyRatesService).to receive(:new)
         allow(ExchangeRates::UploadMonthlyFileService).to receive(:call)
         allow(SlackNotifierService).to receive(:call)
-        allow(ExchangeRates::Mailer).to receive(:monthly_files)
+        allow(ExchangeRatesMailer).to receive(:monthly_files)
 
         call_service
       end
@@ -45,7 +45,7 @@ RSpec.describe ExchangeRates::MonthlyExchangeRatesService do
       it { expect(ExchangeRates::UpdateCurrencyRatesService).not_to have_received(:new) }
       it { expect(ExchangeRates::UploadMonthlyFileService).not_to have_received(:call) }
       it { expect(SlackNotifierService).not_to have_received(:call) }
-      it { expect(ExchangeRates::Mailer).not_to have_received(:monthly_files) }
+      it { expect(ExchangeRatesMailer).not_to have_received(:monthly_files) }
     end
   end
 end
